+++ src/layout/PostAuthLayout/Header/Header.tsx
@@ -5,10 +5,10 @@ import { Menu as AntMenu } from 'antd';
 import { Breadcrumb, Layout, theme } from 'antd';
 import { Link } from '../../../core/common';
 import "./Header.css";
-import { FW24Config } from '../../../core';
+import { UI24Config } from '../../../core';
 import { HeaderActions } from './HeaderActions';
 import { getMethod } from '../../../core';
-import { apiResponse } from '../../../core';
+import { mockApiResponse } from '../../../core';
 
 const { Header : AntHeader } = Layout;
 
@@ -37,48 +37,34 @@ const sampleItems: MenuProps['items'] = [
   ])
 ];
 
-const convertMenuItemsFromApi = (menuItems: any) => {
+
+const formatMenuItems = (menuItems: any) => {
     const items: MenuItem[] = [];
     menuItems.forEach((item: any) => {
         if (item.children) {
-            items.push(createMenuItem(item.label, item.key, <Icon iconName={item.icon} />, convertMenuItemsFromApi(item.children)));
+            items.push(createMenuItem(item.label, item.key, <Icon iconName={item.icon} />, formatMenuItems(item.children)));
         } else {
             items.push(createMenuItem(item.url ? <Link title={item.label} url={item.url} /> : item.label, item.key, <Icon iconName={item.icon} />));
         }
     });
     return items;
-
 }
 
 export const Header = () => {
 
-  const [ menuItems, setMenuItems ] = React.useState<MenuItem[]>( FW24Config.menuItems ? convertMenuItemsFromApi(FW24Config.menuItems) : [] );
-
-  //Load menu from API if menuApiUrl is provided
-  useEffect(()=>{
-    const getMenuFromApi = async () => {
-      const response = await getMethod( FW24Config.menuApiUrl )
-      if( !response ){
-        const { data } = apiResponse( FW24Config.menuApiUrl )
-        setMenuItems( convertMenuItemsFromApi( data ) )
-      }
-    }
-    if( FW24Config.menuApiUrl !== "" ){
-      getMenuFromApi()
-    }
-  },[])
+    const [ menuItems, setMenuItems ] = React.useState<MenuItem[]>( formatMenuItems( UI24Config.uiConfig.menu ?? [] ) );
 
     return <AntHeader style={{ display: 'flex', background: 'white', alignItems: 'center' }}>
       <div className="appHeader">
         <div className="appLogo">
-          { FW24Config?.appLogo !== "" && <div className="logo"><img src={FW24Config.appLogo} alt="App Logo" title="Logo" /></div> }
+          { UI24Config?.appLogo !== "" && <div className="logo"><img src={UI24Config.appLogo} alt="App Logo" title="Logo" /></div> }
         </div>
         <div className="appMenu">
             <AntMenu
               style={{ width: "100%" }}
               theme="light"
               mode="horizontal"
-              items={ menuItems.length > 0 ? menuItems : ( FW24Config.menuApiUrl === "" ? sampleItems : []) }
+              items={ menuItems.length > 0 ? menuItems : ( UI24Config.uiConfig.menu === "" ? sampleItems : []) }
           />
         </div>
         <div className="appActions">