+++ src/layout/PostAuthLayout/PostAuthLayout.tsx
@@ -2,24 +2,26 @@
 //import { lazyLoadComponent } from '../utils/lazyLoadComponent';
 import React, { ReactNode } from 'react';
 import { Layout } from 'antd';
+import { CoreLayout } from '../CoreLayout';
 import { Header } from '../../pages';
-
+import { useAuth } from '../../core';
+import { AppNavigator } from '../../routes/AppRouter';
 
 const { Content, Footer, Sider } = Layout;
-
-const items = new Array(8).fill(null).map((_, index) => ({
-  key: index + 1,
-  label: `nav ${index + 1}`,
-}));
 
 interface IChildren{
     children: ReactNode;
 }
 
 export const PostAuthLayout: React.FC<IChildren> = ({ children }) => {
+  const { isLoggedIn } = useAuth();
+
+  if( !isLoggedIn ) {
+    return <AppNavigator />
+}
   
   return (
-    <Layout style={{ minHeight: "100vh"}}>
+    <CoreLayout >
       <Header />
       <Layout >
         <Content >
@@ -29,6 +31,6 @@ export const PostAuthLayout: React.FC<IChildren> = ({ children }) => {
       <Footer style={{ textAlign: 'center' }}>
         Ten24 ©{new Date().getFullYear()}
       </Footer>
-    </Layout>
+    </CoreLayout>
   );
 };