+++ src/pages/FW24.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-import { AppRouter, IAppRouter } from '../routes/AppRouter';
-
-const FW24 = ({ customRoutes = [] }: IAppRouter ) => {
-    return <AppRouter customRoutes = { customRoutes } />
-}
-export { FW24 };