+++ src/pages/PostAuth/PostAuthPage.tsx
@@ -3,12 +3,14 @@ import { PostAuthLayout } from "../../layout"
 import { PageHeader, IPageHeader } from './PageHeader/PageHeader';
 import { ICustomForm } from '../../core/forms/formConfig';
 import "./PostAuthPage.css";
-import { Card, Alert } from 'antd';
+import { Card } from 'antd';
 import { PostAuthForm } from '../../forms/PostAuthForm';
-import { Table, ITableConfig } from '../../table/Table';
+import { Table } from '../../table/Table';
+import { ITableConfig } from '../../table/type';
 import { Details, IDetailsConfig } from '../../detail/Details';
+import { v4 as uuidv4 } from 'uuid';
 
-type IPageType = "list" | "form" | "accordion" | "details | dashboard"
+type IPageType = "list" | "form" | "accordion" | "details" | "dashboard";
 
 interface IRenderFromPageType {
     pageType?: IPageType;
@@ -19,13 +21,12 @@ interface IRenderFromPageType {
 }
 
 export interface IPostAuthPage extends IPageHeader, IRenderFromPageType {
-    metaDataUrl: string;
     CustomPageHeader?: React.ReactNode;
     children?: React.ReactNode;
 }
 
 
-export const PostAuthPage = ({ metaDataUrl, CustomPageHeader, children, ...props } : IPostAuthPage ) => {
+export const PostAuthPage = ({ CustomPageHeader, children, ...props } : IPostAuthPage ) => {
     return <PostAuthLayout>
         <div style={{ paddingTop: "1%"}}>
             <div className = "PostAuthContainer" >
@@ -42,7 +43,7 @@ export const PostAuthPage = ({ metaDataUrl, CustomPageHeader, children, ...props
 export const RenderFromPageType = ( {pageType, cardStyle, formPageConfig, listPageConfig, detailsPageConfig}: IRenderFromPageType ) => {
     
     switch( pageType ) {
-        case "list": return <Card style={ cardStyle } > <Table {...listPageConfig} /> </Card>;
+        case "list": return <Card style={ cardStyle } > <Table {...listPageConfig} key={`list-${uuidv4()}`} /> </Card>;
         case "form": return <Card style={ cardStyle } > <PostAuthForm {...formPageConfig} /> </Card>;
         case "details": return <Card style={ cardStyle } > <Details {...detailsPageConfig} /> </Card>;
         case "accordion": return <div> Accordion Page </div>;