+++ src/pages/index.tsx
@@ -1,5 +1,4 @@
-export { FW24 } from './FW24'
-
+import React from 'react';
 //Post Login pages
 export { PostAuthPage } from "./PostAuth/PostAuthPage"
 export { Header } from '../layout/PostAuthLayout/Header/Header'
@@ -7,5 +6,7 @@ export { DynamicPage } from "./DynamicPage/DynamicPage"
 
 //Pre Login Pages
 export { LoginPage } from "./PreAuth/LoginPage"
-export { ForgotPassword } from './PreAuth/ForgotPassword'
-export { ResetPassword } from './PreAuth/ResetPassword'
+export { ForgotPasswordPage } from './PreAuth/ForgotPassword'
+export { ResetPasswordPage } from './PreAuth/ResetPassword'
+export { RegistrationPage } from './PreAuth/RegistrationPage'
+export { VerifyRegistrationPage } from './PreAuth/VerifyRegistrationPage';