+++ CHANGELOG.md
@@ -2,6 +2,48 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [0.0.15](https://github.com/ten24group/ui24/compare/v0.0.14...v0.0.15) (2024-05-15)
+
+### [0.0.14](https://github.com/ten24group/ui24/compare/v0.0.13...v0.0.14) (2024-05-15)
+
+### [0.0.13](https://github.com/ten24group/ui24/compare/v0.0.12...v0.0.13) (2024-05-15)
+
+
+### Bug Fixes
+
+* app-context type errors ([c945566](https://github.com/ten24group/ui24/commit/c9455669968df2c19f07b39f6d2adc36772a6303))
+* login, forgot-password and reset-password pages ([92e7947](https://github.com/ten24group/ui24/commit/92e794798efb93f532d176d8767d028042cde9d8))
+* notofications not working on navigate ([9eb91ff](https://github.com/ten24group/ui24/commit/9eb91ffdc5fd21bd254cd477572d118a547d33a8))
+* static error message on API errors ([eebf191](https://github.com/ten24group/ui24/commit/eebf1915eb4967ae776d6e3f473a5316d25bed6b))
+* type-errors in Button.tsx and add leftover file for logout implementation ([001c631](https://github.com/ten24group/ui24/commit/001c6318348fd1b2bcbb75ce690b073e88c40033))
+
+### [0.0.12](https://github.com/ten24group/ui24/compare/v0.0.11...v0.0.12) (2024-05-09)
+
+### [0.0.11](https://github.com/ten24group/ui24/compare/v0.0.10...v0.0.11) (2024-05-08)
+
+
+### Bug Fixes
+
+* aws signature issue due to extra slash ([24027cc](https://github.com/ten24group/ui24/commit/24027cc1298f3caa94d02bb64ff66e063fdf15cb))
+
+### [0.0.10](https://github.com/ten24group/ui24/compare/v0.0.9...v0.0.10) (2024-05-08)
+
+
+### Bug Fixes
+
+* type in axios base-url env var ([69ae7a4](https://github.com/ten24group/ui24/commit/69ae7a480b843aa1dd1c3d703526476622ebe5c7))
+
+### [0.0.9](https://github.com/ten24group/ui24/compare/v0.0.8...v0.0.9) (2024-05-08)
+
+### [0.0.8](https://github.com/ten24group/ui24/compare/v0.0.7...v0.0.8) (2024-05-07)
+
+
+### Bug Fixes
+
+* dashboard navigation ([fdaf5f6](https://github.com/ten24group/ui24/commit/fdaf5f6d0fd87677ed06aee9631beb4d5e120921))
+* login not working ([538453a](https://github.com/ten24group/ui24/commit/538453adaec774d29c2c57f3defae191f9ad8128))
+* redirect to dashboard on successful login ([ca08139](https://github.com/ten24group/ui24/commit/ca081398bf27e6d577f4e560fd3d45095bcf9f62))
+
 ### [0.0.7](https://github.com/ten24group/ui24/compare/v0.0.6...v0.0.7) (2024-03-22)
 
 ### [0.0.6](https://github.com/ten24group/ui24/compare/v0.0.5...v0.0.6) (2024-03-21)